fix: signedness comparison in IrtInterfaceJsonParser
/home/runner/work/EICrecon/EICrecon/src/algorithms/pid/IrtInterfaceJsonParser.cc: In member function 'void eicrecon::IrtInterface::JsonParser()':
/home/runner/work/EICrecon/EICrecon/src/algorithms/pid/IrtInterfaceJsonParser.cc:104:26: error: comparison of integer expressions of different signedness: 'int' and 'std::map<TString, IRT2::CherenkovRadiator*>::size_type' {aka 'long unsigned int'} [-Werror=sign-compare]
  104 |                 if (rnum == reco->GetMyRICH()->Radiators().size()) {
      |                     ~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/runner/work/EICrecon/EICrecon/src/algorithms/pid/IrtInterfaceJsonParser.cc:110:44: error: comparison of integer expressions of different signedness: 'unsigned int' and 'int' [-Werror=sign-compare]
  110 |                   for (unsigned ir = 0; ir < rnum; ir++)
      |                                         ~~~^~~~~~

+++ src/algorithms/pid/IrtInterfaceJsonParser.cc
@@ -99,7 +99,7 @@ void IrtInterface::JsonParser(void) {
 
                 auto* calib = &radiator->m_Calibrations[iq];
 
-                int rnum = (int)tarray.size() - 4;
+                unsigned rnum = tarray.size() - 4;
 
                 if (rnum == reco->GetMyRICH()->Radiators().size()) {
                   calib->m_Stat        = atoi(tarray[0].template get<std::string>().c_str());
